+++ src/main/java/insane96mcp/enhancedai/mixin/RavagerMixin.java
@@ -1,22 +1,24 @@
 package insane96mcp.enhancedai.mixin;
 
+import com.llamalad7.mixinextras.expression.Definition;
+import com.llamalad7.mixinextras.expression.Expression;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
+import com.llamalad7.mixinextras.sugar.Local;
 import insane96mcp.enhancedai.modules.illager.RavagerFeature;
 import net.minecraft.world.entity.monster.Ravager;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.LeavesBlock;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.Constant;
-import org.spongepowered.asm.mixin.injection.ModifyConstant;
+import org.spongepowered.asm.mixin.injection.At;
 
 @Mixin(Ravager.class)
 public class RavagerMixin {
-	@ModifyConstant(method = "aiStep()V", constant = @Constant(classValue = LeavesBlock.class))
-	public boolean aiStep_onInstanceOf(Object targetObj, Class<?> classValue) {
-		return targetObj != null && ((Block)targetObj).builtInRegistryHolder().is(RavagerFeature.BREAKABLE_BY_RAVAGER);
+	@Definition(id = "block", local = @Local(type = Block.class))
+	@Definition(id = "LeavesBlock", type = LeavesBlock.class)
+	@Expression("block instanceof LeavesBlock")
+	@WrapOperation(method = "aiStep", at = @At("MIXINEXTRAS:EXPRESSION"))
+	public boolean aiStep_onInstanceOf(Object object, Operation<Boolean> original) {
+		return object != null && ((Block)object).builtInRegistryHolder().is(RavagerFeature.BREAKABLE_BY_RAVAGER);
 	}
-
-	/*@ModifyConstant(method = "aiStep()V", constant = @Constant(classValue = LeavesBlock.class))
-	public Class<?> aiStep_onInstanceOf(Class<?> constant) {
-		return ((Block)targetObj).builtInRegistryHolder().is(RavagerFeature.BREAKABLE_BY_RAVAGER);
-	}*/
 }