+++ src/main/java/insane96mcp/enhancedai/modules/animal/NotTemptedAnimals.java
@@ -4,13 +4,11 @@ import insane96mcp.enhancedai.EnhancedAI;
 import insane96mcp.enhancedai.modules.Modules;
 import insane96mcp.enhancedai.setup.NBTUtils;
 import insane96mcp.insanelib.base.Feature;
-import insane96mcp.insanelib.base.Label;
 import insane96mcp.insanelib.base.LoadFeature;
 import insane96mcp.insanelib.base.Module;
 import insane96mcp.insanelib.base.config.Config;
 import net.minecraft.core.registries.Registries;
 import net.minecraft.nbt.CompoundTag;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.EntityType;
 import net.minecraft.world.entity.ai.goal.TemptGoal;
@@ -20,15 +18,13 @@ import net.minecraftforge.event.entity.EntityJoinLevelEvent;
 import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 
-@Label(name = "Not Tempted Animals", description = "Makes animals not tempted by food. Use the entity type tag enhancedai:can_ignore_food_temptation to change animals.")
-@LoadFeature(module = Modules.Ids.ANIMAL)
+@LoadFeature(module = Modules.Ids.ANIMAL, description = "Makes animals not tempted by food. Use the entity type tag enhancedai:can_ignore_food_temptation to change animals.")
 public class NotTemptedAnimals extends Feature {
     //TODO Some animals should attack the player with the food in the hand
-    public static final TagKey<EntityType<?>> CAN_IGNORE_FOOD_TEMPTATION = TagKey.create(Registries.ENTITY_TYPE, new ResourceLocation(EnhancedAI.MOD_ID, "can_ignore_food_temptation"));
+    public static final TagKey<EntityType<?>> CAN_IGNORE_FOOD_TEMPTATION = TagKey.create(Registries.ENTITY_TYPE, EnhancedAI.location("can_ignore_food_temptation"));
     public static final String NOT_TEMPTED = EnhancedAI.RESOURCE_PREFIX + "not_tempted";
 
-    @Config(min = 0d, max = 1d)
-    @Label(name = "Not tempted chance", description = "Animals have this percentage chance to not be temped by food.")
+    @Config(min = 0d, max = 1d, description = "Animals have this percentage chance to not be temped by food.")
     public static Double notTemptedChance = 0.5d;
 
     public NotTemptedAnimals(Module module, boolean enabledByDefault, boolean canBeDisabled) {